feat(gp-dynamic-fees): update formatting

+++ test/GasPriceFeesHook.t.sol
@@ -27,7 +27,7 @@ contract TestGasPriceFeesHook is Test, Deployers {
 
     // In the set up, we deploy v4-core contracts, mint some tokens, deploy
     // our hook, initialize a new pool, and add some liquidity to the pool.
-	function setUp() public {
+    function setUp() public {
         // Deploy v4-core
         deployFreshManagerAndRouters();
 
@@ -35,13 +35,8 @@ contract TestGasPriceFeesHook is Test, Deployers {
         deployMintAndApprove2Currencies();
 
         // Deploy our hook with the proper flags
-        address hookAddress = address(
-            uint160(
-                Hooks.BEFORE_INITIALIZE_FLAG |
-                Hooks.BEFORE_SWAP_FLAG |
-                Hooks.AFTER_SWAP_FLAG
-            )
-        );
+        address hookAddress =
+            address(uint160(Hooks.BEFORE_INITIALIZE_FLAG | Hooks.BEFORE_SWAP_FLAG | Hooks.AFTER_SWAP_FLAG));
 
         // Set gas price = 10 gwei and deploy our hook
         // (the default in Foundry is to use 0 gas price).
@@ -50,7 +45,7 @@ contract TestGasPriceFeesHook is Test, Deployers {
         hook = GasPriceFeesHook(hookAddress);
 
         // Initialize a pool
-        (key, ) = initPool(
+        (key,) = initPool(
             currency0,
             currency1,
             hook,
@@ -69,7 +64,7 @@ contract TestGasPriceFeesHook is Test, Deployers {
             }),
             ZERO_BYTES
         );
-	}
+    }
 
     // Test scenario:
     // 1. Ensure that the current moving average gas price is set to 10 gwei - as that is what was
@@ -84,8 +79,8 @@ contract TestGasPriceFeesHook is Test, Deployers {
     // between the output amount of tokens we get back from the swap isn't because of a price shift.
     function test_feeUpdatesWithGasPrice() public {
         // Set up swap parameters
-        PoolSwapTest.TestSettings memory testSettings = PoolSwapTest
-            .TestSettings({takeClaims: false, settleUsingBurn: false});
+        PoolSwapTest.TestSettings memory testSettings =
+            PoolSwapTest.TestSettings({takeClaims: false, settleUsingBurn: false});
 
         IPoolManager.SwapParams memory params = IPoolManager.SwapParams({
             zeroForOne: true,
@@ -108,8 +103,7 @@ contract TestGasPriceFeesHook is Test, Deployers {
         uint256 balanceOfToken1Before = currency1.balanceOfSelf();
         swapRouter.swap(key, params, testSettings, ZERO_BYTES);
         uint256 balanceOfToken1After = currency1.balanceOfSelf();
-        uint256 outputFromBaseFeeSwap = balanceOfToken1After -
-            balanceOfToken1Before;
+        uint256 outputFromBaseFeeSwap = balanceOfToken1After - balanceOfToken1Before;
 
         assertGt(balanceOfToken1After, balanceOfToken1Before);
 
@@ -127,8 +121,7 @@ contract TestGasPriceFeesHook is Test, Deployers {
         swapRouter.swap(key, params, testSettings, ZERO_BYTES);
         balanceOfToken1After = currency1.balanceOfSelf();
 
-        uint256 outputFromIncreasedFeeSwap = balanceOfToken1After -
-            balanceOfToken1Before;
+        uint256 outputFromIncreasedFeeSwap = balanceOfToken1After - balanceOfToken1Before;
 
         assertGt(balanceOfToken1After, balanceOfToken1Before);
 
@@ -145,8 +138,7 @@ contract TestGasPriceFeesHook is Test, Deployers {
         swapRouter.swap(key, params, testSettings, ZERO_BYTES);
         balanceOfToken1After = currency1.balanceOfSelf();
 
-        uint outputFromDecreasedFeeSwap = balanceOfToken1After -
-            balanceOfToken1Before;
+        uint256 outputFromDecreasedFeeSwap = balanceOfToken1After - balanceOfToken1Before;
 
         assertGt(balanceOfToken1After, balanceOfToken1Before);
 
@@ -156,7 +148,6 @@ contract TestGasPriceFeesHook is Test, Deployers {
 
         assertEq(movingAverageGasPrice, 9 gwei);
         assertEq(movingAverageGasPriceCount, 4);
-
 
         // 4. Check all the output amounts
         console.log("Base Fee Output", outputFromBaseFeeSwap);
